pyproject.toml
[tool.poetry]
name = "test-task-insiders"
version = "0.1.0"
description = ""
authors = ["DaTrEvTeR <rozhkovm176@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.7"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-filter = "^25.1"
drf-spectacular = "^0.28.0"
redis = "^5.2.1"
django-environ = "^0.12.0"
pandas = "^2.2.3"
drf-nested-routers = "^0.94.1"
django-redis = "^5.4.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
ignore = ["F401"]
